refactor, add test browser env variable
