+++ Jenkinsfile
@@ -176,6 +176,7 @@ ZOWE_USERNAME=${USERNAME} \
 ZOWE_PASSWORD=${PASSWORD} \
 SERVER_HOST_NAME=${params.FVT_SERVER_HOSTNAME} \
 SERVER_HTTPS_PORT=7554 \
+TEST_BROWSER=firefox \
 npm run test:fvt
 """
           }
