+++ package.json
@@ -59,11 +59,10 @@
     "eslint-plugin-node": "^5.1.1",
     "eslint-plugin-react": "^7.3.0",
     "expect": "^1.20.2",
-    "explorer-fvt-utilities": "1.0.4",
-    "geckodriver": "^1.19.1",
+    "explorer-fvt-utilities": "1.0.6-pr-3-11-20200611051521",
     "isomorphic-fetch": "^2.2.1",
-    "mocha": "^6.1.4",
-    "mocha-jenkins-reporter": "^0.4.1",
+    "mocha": "^7.0.1",
+    "mocha-jenkins-reporter": "^0.4.2",
     "nock": "^9.6.1",
     "node-fetch": "^2.6.0",
     "npm-run-all": "^4.1.2",
