+++ tests/FVTTests/README.md
@@ -30,5 +30,6 @@ ZOWE_USERNAME=IBMUSER \
 ZOWE_PASSWORD=12345678 \
 SERVER_HOST_NAME=zosHost.com \
 SERVER_HTTPS_PORT=7554 \
+TEST_BROWSER=firefox \
 npm run test:fvt
 ```
