+++ tests/FVTTests/filterParametersView/testFilterParameterView.ts
@@ -14,11 +14,8 @@ import { expect } from 'chai';
 
 const chai = require('chai');
 chai.use(require('chai-things'));
-require('geckodriver');
 
 import {
-    getDriver,
-    setApimlAuthTokenCookie,
     testElementAppearsXTimesByCSS 
 } from 'explorer-fvt-utilities';
 
@@ -37,15 +34,11 @@ import {
     testFilterFormInputValues,
 } from '../testFunctions';
 
-const {
-    ZOWE_USERNAME: USERNAME, ZOWE_PASSWORD: PASSWORD, SERVER_HOST_NAME, SERVER_HTTPS_PORT,
-} = process.env;
+import {getDriver, TEST_CONFIG} from "../../FVTTests/testConfig";
 
-const BASE_URL = `https://${SERVER_HOST_NAME}:${SERVER_HTTPS_PORT}`;
-const BASE_URL_WITH_PATH = `${BASE_URL}/ui/v1/explorer-jes`;
-const FILTER_BASE_URL = `${BASE_URL_WITH_PATH}/#/`;
-const loadUrlWithSearchFilters = loadPageWithFilterOptions(FILTER_BASE_URL, DEFAULT_SEARCH_FILTERS);
-const ZOSMF_JOB_NAME = 'IZUSVR1';
+const {
+    FILTER_BASE_URL, loadUrlWithSearchFilters, ZOSMF_JOB_NAME
+} = TEST_CONFIG;
 
 // Need to use unnamed function so we can specify the retries
 // eslint-disable-next-line
@@ -55,7 +48,7 @@ describe('JES explorer home view with filter parameters in url query', function 
 
     before('Initialise', async () => {
         driver = await getDriver();
-        await setApimlAuthTokenCookie(driver, USERNAME, PASSWORD, `${BASE_URL}/api/v1/gateway/auth/login`, BASE_URL_WITH_PATH);
+        // await setApimlAuthTokenCookie(driver, USERNAME, PASSWORD, `${BASE_URL}/api/v1/gateway/auth/login`, BASE_URL_WITH_PATH);
     });
 
     after('Close out', async () => {
